refactor: TodoList, ScheduleList component/common으로 이동
