+++ src/pages/main/MainPage.tsx
@@ -1,6 +1,7 @@
 import Calendar from "@/components/common/Calendar/Calendar";
-import ScheduleList from "@/components/feature/main/ScheduleList/ScheduleList";
 import CategoryButtons from "@/components/feature/main/CategoryButtons/CategoryButtons";
+import ScheduleList from "@/components/common/ScheduleList/ScheduleList";
+import TodoList from "@/components/common/TodoList/TodoList";
 
 function MainPage() {
   return (
@@ -8,6 +9,7 @@ function MainPage() {
       <Calendar usage="main" onDateSelect={() => {}} />
       <CategoryButtons />
       <ScheduleList />
+      <TodoList />
     </>
   );
 }
